restructure the source code directory
