+++ dockerfiles/alpine-glibc.dockerfile
@@ -5,10 +5,10 @@ RUN apk add --no-cache pkgconf binutils python3 py3-pip
 RUN pip install --no-cache pyinstaller pillow python-docx mysql-connector-python click
 COPY rekdoc/ /package/rekdoc/
 RUN pyinstaller --strip --clean \
-    -F rekdoc/core.py rekdoc/doc.py rekdoc/fetch.py rekdoc/const.py rekdoc/tools.py -n rekdoc
+    -F rekdoc/core.py rekdoc/doc.py rekdoc/fetch.py rekdoc/const.py rekdoc/tools.py -n rd
 
 FROM jeanblanchard/alpine-glibc:3.18 as final
-COPY --from=base /package/dist/rekdoc /usr/bin/rekdoc 
+COPY --from=base /package/dist/rd /usr/bin/rd
 RUN adduser -D py
 WORKDIR /home/py
 USER py
