+++ dockerfiles/debian.dockerfile
@@ -13,10 +13,10 @@ COPY rekdoc/ /package/rekdoc
 RUN pyinstaller --strip --clean \
     -F rekdoc/core.py rekdoc/doc.py rekdoc/fetch.py rekdoc/const.py rekdoc/tools.py \
 		--add-binary rekdoc/oswbba.jar:. \
-    -n rekdoc
+    -n rd
 
 FROM debian:buster-slim as final
-COPY --from=base /package/dist/rekdoc /usr/bin/rekdoc 
+COPY --from=base /package/dist/rd /usr/bin/rd
 RUN useradd py -u 1000
 USER py
 WORKDIR /home/py
