+++ pyproject.toml
@@ -24,7 +24,7 @@ requires-python = ">=3"
 readme = "README.md"
 
 [project.scripts]
-rekdoc = "rekdoc.core:cli"
+rd = "rekdoc.core:cli"
 
 [project.optional-dependencies]
 dev = ["hatch",
